+++ CHANGELOG.md
@@ -23,6 +23,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+* Fix R&F optimization converters to use optimized channels.
 * Fix a JAX compatibility issue in `Summarizer` when calling `.idxmax()` on data
   backed by JAX.
 * Add support for posterior downsampling for faster, near approximate inference.